FullRequestController

+++ src/main/java/com/totoApp/model/Customer.java
@@ -8,7 +8,7 @@ public class Customer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     private String firstName;
     private String lastName;
@@ -22,6 +22,8 @@ public class Customer {
     public Customer() {
 
     }
+
+
 
     public String getFirstName() {
         return firstName;
